backend/cdn/requirements.txt
Pillow==10.2.0
requests==2.31.0











